Add initial magic bitboard implementation.
This is ruinously slow in debug mode (~20s to init on a strong
workstation. Release mode makes a big difference (<3s), but I need to
investigate some ways of speeding up the debug build so I can actually
use it in development and the test cycle doesn't make me crazy.