Movegen performance optimizations.
This puts the rust implementation on par with my C++ equivalent. The
biggest by far is to pre-allocate space for moves, but the color
flipping change is worth a few percentage points. I also experimented
with inlining and get_unchecked, but these didn't deliver noticeable
improvements so I ditched them.