update move selection phase logic
previously, the selection phase was updated after generation and
ordering, so when you actually chose the move, the 'current' phase was
actually the next phase (i.e. the phase that will generate the next
batch of moves). This leads to all the phase-specific behaviors inside
the move-returning part of next() being wrong.